extracting sampler agg
